requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
coverage==7.2.3
Django==4.2
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.4
